+++ crates/store/src/db/schema.rs
@@ -1,51 +1,6 @@
 // @generated automatically by Diesel CLI.
 
 diesel::table! {
-    account_deltas (account_id, block_num) {
-        account_id -> Binary,
-        block_num -> BigInt,
-        nonce -> BigInt,
-    }
-}
-
-diesel::table! {
-    account_fungible_asset_deltas (account_id, block_num, faucet_id) {
-        account_id -> Binary,
-        block_num -> BigInt,
-        faucet_id -> Binary,
-        delta -> BigInt,
-    }
-}
-
-diesel::table! {
-    account_non_fungible_asset_updates (account_id, block_num, vault_key) {
-        account_id -> Binary,
-        block_num -> BigInt,
-        vault_key -> Binary,
-        is_remove -> Bool,
-    }
-}
-
-diesel::table! {
-    account_storage_map_updates (account_id, block_num, slot, key) {
-        account_id -> Binary,
-        block_num -> BigInt,
-        slot -> Integer,
-        key -> Binary,
-        value -> Binary,
-    }
-}
-
-diesel::table! {
-    account_storage_slot_updates (account_id, block_num, slot) {
-        account_id -> Binary,
-        block_num -> BigInt,
-        slot -> Integer,
-        value -> Binary,
-    }
-}
-
-diesel::table! {
     accounts (account_id) {
         account_id -> Binary,
         network_account_id_prefix -> Nullable<BigInt>,
@@ -117,8 +72,6 @@ diesel::table! {
     }
 }
 
-diesel::joinable!(account_deltas -> accounts (account_id));
-diesel::joinable!(account_deltas -> block_headers (block_num));
 diesel::joinable!(accounts -> account_codes (code_commitment));
 diesel::joinable!(accounts -> block_headers (block_num));
 diesel::joinable!(notes -> accounts (sender));
@@ -129,12 +82,7 @@ diesel::joinable!(transactions -> accounts (account_id));
 diesel::joinable!(transactions -> block_headers (block_num));
 
 diesel::allow_tables_to_appear_in_same_query!(
-    account_deltas,
     account_codes,
-    account_fungible_asset_deltas,
-    account_non_fungible_asset_updates,
-    account_storage_map_updates,
-    account_storage_slot_updates,
     accounts,
     block_headers,
     note_scripts,